Tidy config merge review follow-ups
Move shared CLI merge test helpers into the shared helper module
so the main merge test file stays below the repository line limit.
Merge the duplicate `TestEnv` impl and fix the developer-guide grammar.
